feat: 비즈니스 로직을 포함한 MojiBookPage 생성 및 컴포넌트 구조 개선
