+++ src/app/page.tsx
@@ -1,11 +1,9 @@
-'use client';
-
-import InteractiveBook from '@/component/InteractiveBook';
+import MojiBookPage from '@/containers/MojiBookPage';
 
 export default function Home() {
     return (
         <div className="flex items-center justify-center h-screen">
-            <InteractiveBook />
+            <MojiBookPage />
         </div>
     );
 }
