+++ src/components/FlippingBook/index.tsx
@@ -1,20 +1,23 @@
 'use client';
 import { motion } from 'framer-motion';
-import BookCoverInner from '@/component/Book/BookCoverInner';
 import { useCallback, useMemo, useState } from 'react';
-import useMojiMessage from '@/app/hooks/api/useMojiMessage';
-import BookPages from '@/component/Book/BookPages';
-import BookZoomed from '@/component/Book/BookZoomed';
+import ZoomPage from '@/components/FlippingBook/page/ZoomPage';
+import Pages from './page/Pages';
+import BookCoverInner from '../common/Book/BookCoverInner';
 
-const FlippingBook = () => {
+interface FlippingBookProps {
+    getMojiMessage: () => void;
+    message: string;
+}
+
+const FlippingBook = ({ getMojiMessage, message }: FlippingBookProps) => {
     const [currentStep, setCurrentStep] = useState<'closed' | 'zoomed'>('closed');
-    const { mutateAsync: getMojiMessage, data: message } = useMojiMessage();
     const pages = useMemo(() => Array(8).fill(0), []);
 
     const handleAnimationComplete = useCallback(
         async (index: number) => {
             if (index !== pages.length - 1) return;
-            // await getMojiMessage();
+            await getMojiMessage();
             setCurrentStep('zoomed');
         },
         [getMojiMessage],
@@ -35,8 +38,8 @@ const FlippingBook = () => {
 
     const viewConfig = useMemo(
         () => ({
-            closed: <BookPages {...propsConfig['closed']} />,
-            zoomed: <BookZoomed {...propsConfig['zoomed']} />,
+            closed: <Pages {...propsConfig['closed']} />,
+            zoomed: <ZoomPage {...propsConfig['zoomed']} />,
         }),
         [propsConfig],
     );
