+++ src/components/FlippingBook/page/ZoomPage.tsx
@@ -1,10 +1,10 @@
 import { motion } from 'framer-motion';
 
-interface BookZoomedProps {
+interface ZoomPageProps {
     message: string;
 }
 
-const BookZoomed = ({ message }: BookZoomedProps) => {
+const ZoomPage = ({ message }: ZoomPageProps) => {
     return (
         <motion.div
             className="absolute top-0 left-0 w-full h-full bg-white box-shadow-lg p-40 z-100 flex items-center justify-center transform-origin-center pointer-events-auto rounded-sm"
@@ -20,4 +20,4 @@ const BookZoomed = ({ message }: BookZoomedProps) => {
     );
 };
 
-export default BookZoomed;
+export default ZoomPage;
