+++ src/components/common/Book/BookCoverInner.tsx
